pyproject.toml
[project]
name = "ai-hr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.30.0",
    "chainlit>=2.8.0",
    "fastapi>=0.116.2",
    "logfire>=4.8.0",
    "pydantic>=2.11.9",
    "pydantic-ai-slim[openai]>=1.0.10",
    "psycopg2-binary>=2.9.0",
    "python-dotenv>=1.0.0",
    "sqlalchemy>=2.0.43",
]
